+++ src/main/java/com/apap/TAsilab/repository/PemeriksaanDB.java
@@ -3,6 +3,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.apap.TAsilab.model.*;
 
-public interface PemeriksaanDB extends JpaRepository<PemeriksaanModel, Long>  {
+public interface PemeriksaanDB extends JpaRepository<PemeriksaanModel, Integer>  {
 	
 }
