+++ src/main/java/com/apap/TAsilab/service/PemeriksaanService.java
@@ -7,7 +7,7 @@ import com.apap.TAsilab.model.PemeriksaanModel;
 
 
 public interface PemeriksaanService {
-	PemeriksaanModel findPemeriksaanById(long id);
+	PemeriksaanModel findPemeriksaanById(int id);
 	List<PemeriksaanModel> findAll();
 	void updatePemeriksaan(PemeriksaanModel pemeriksaan);
